Buttons & Update Checking
If statements allow articles and events to be checked as the wether
they have been updated or not. Could be DRYer but the helper class
doesnt work correctly. - Working!

All buttons are there for editing articles and events. All pages are
linked up and working.

Only known problem is that when a article/event is edited relative
links wont work as we are too far up he hierarchy.